Apply refactor
