+++ gnd/src/main/java/com/google/android/gnd/ui/home/mapcontainer/PolygonDrawingView.java
@@ -25,11 +25,9 @@ import com.google.android.gnd.R;
 import com.google.android.gnd.databinding.PolygonDrawingControlsBinding;
 import com.google.android.gnd.ui.common.AbstractView;
 import com.google.android.gnd.ui.map.CameraPosition;
-import com.google.android.gnd.ui.map.MapAdapter;
-import com.google.android.gnd.ui.map.MapProvider;
+import com.google.android.gnd.ui.map.MapFragment;
 import dagger.hilt.android.AndroidEntryPoint;
 import dagger.hilt.android.WithFragmentBindings;
-import javax.inject.Inject;
 
 @WithFragmentBindings
 @AndroidEntryPoint
@@ -37,7 +35,6 @@ public class PolygonDrawingView extends AbstractView {
 
   private final PolygonDrawingViewModel viewModel;
   private final PolygonDrawingControlsBinding binding;
-  @Inject MapProvider mapProvider;
 
   public PolygonDrawingView(Context context, @Nullable AttributeSet attrs) {
     super(context, attrs);
@@ -47,12 +44,9 @@ public class PolygonDrawingView extends AbstractView {
     binding.setViewModel(viewModel);
   }
 
-  @Override
-  protected void onFinishInflate() {
-    super.onFinishInflate();
-    mapProvider
-        .getMapAdapter()
-        .flatMap(MapAdapter::getCameraMovedEvents)
+  public void init(MapFragment mapFragment) {
+    mapFragment
+        .getCameraMovedEvents()
         .map(CameraPosition::getTarget)
         .onBackpressureLatest()
         .as(disposeOnDestroy(getActivity()))
