+++ gnd/src/main/java/com/google/android/gnd/ui/map/MapFragmentFactory.java
@@ -1,0 +1,11 @@
+package com.google.android.gnd.ui.map;
+
+import androidx.fragment.app.Fragment;
+import com.google.common.collect.ImmutableList;
+
+public interface MapFragmentFactory {
+
+  Fragment createFragment();
+
+  ImmutableList<MapType> getMapTypes();
+}
